agent/requirements.txt
google-adk==1.9.0
nest_asyncio
toolbox-core
a2a-sdk==0.2.16